requirements.txt
mysql-connector-python
paho-mqtt
django
djangorestframework
markdown
django-filter
mysqlclient
